Replace ClientError when trying to accept an expired invite with a better message
ClientErrors signal when the client has a bug, or is otherwise violating 
the protocol. However, failing to accept an expired invite, is something 
that the client can't reasonably prevent, so it must be handled by the 
server. Ideally ClientError messages will never be seen by legitimate 
players as their client implementations should be correct. However, this 
message likely will be seen, so it should be passed in a way that it can 
be locallized into the player's language of choice.